+++ src/App.jsx
@@ -1,33 +1,13 @@
 import "./App.css";
-
-const todoList = [
-  {
-    id: 1,
-    title: "Grocery store shopping",
-  },
-  {
-    id: 2,
-    title: "Do laundry",
-  },
-  {
-    id: 3,
-    title: "Complete assignments",
-  },
-  {
-    id: 4,
-    title: "Practice Leetcode problems",
-  },
-];
+import TodoList from "./TodoList";
+import AddTodoForm from "./AddTodoForm";
 
 function App() {
   return (
     <div>
       <h1>Todo List</h1>
-      <ul>
-        {todoList.map((item) => {
-          return <li key={item.id}>{item.title}</li>;
-        })}
-      </ul>
+      <AddTodoForm />
+      <TodoList />
     </div>
   );
 }
